사용자는 홈에 들어가면 인기 모임리스트부터 본다 화면이 처음 렌더링되면 인기 모임리스트부터 보이게 로직을 작성했습니다. 커스텀훅 파일명을 변경했습니다.
